chore: adopt connector-data, remove connector-source/destination
