+++ config/config.yaml
@@ -28,7 +28,7 @@ database:
   host: pg-sql
   port: 5432
   name: connector
-  version: 4
+  version: 5
   timezone: Etc/UTC
   pool:
     idleconnections: 5
