+++ integration-test/proto/vdp/connector/v1alpha/connector_definition.proto
@@ -2,14 +2,12 @@ syntax = "proto3";
 
 package vdp.connector.v1alpha;
 
+import "google/api/field_behavior.proto";
+// Google API
+import "google/api/resource.proto";
 // Protocol Buffers Well-Known Types
 import "google/protobuf/struct.proto";
 import "protoc-gen-openapiv2/options/annotations.proto";
-
-// Google API
-import "google/api/resource.proto";
-import "google/api/field_behavior.proto";
-
 import "../../../vdp/connector/v1alpha/spec.proto";
 
 // View enumerates the definition views
@@ -36,53 +34,58 @@ enum ConnectorType {
   CONNECTOR_TYPE_AI = 3;
   // ConnectorType: Blockchain
   CONNECTOR_TYPE_BLOCKCHAIN = 4;
+  // ConnectorType: DATA
+  CONNECTOR_TYPE_DATA = 5;
+  // ConnectorType: OPERATOR
+  CONNECTOR_TYPE_OPERATOR = 6;
 }
 
 // ConnectorDefinition represents the connector definition data model
 message ConnectorDefinition {
   option (google.api.resource) = {
-    type : "api.instill.tech/ConnectorDefinition"
-    pattern : "connector-definitions/"
-              "{connector-definition}"
+    type: "api.instill.tech/ConnectorDefinition"
+    pattern:
+      "connector-definitions/"
+      "{connector-definition}"
   };
 
   // ConnectorDefinition resource name. It must have the format of
   // "connector-definitions/*"
-  string name = 1 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  string name = 1 [(google.api.field_behavior) = OUTPUT_ONLY];
   // ConnectorDefinition UUID
-  string uid = 2 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  string uid = 2 [(google.api.field_behavior) = OUTPUT_ONLY];
   // ConnectorDefinition resource ID (the last segment of the
   // resource name) used to construct the resource name. This conforms to
   // RFC-1034, which restricts to letters, numbers, and hyphen, with the first
   // character a letter, the last a letter or a number, and a 63 character
   // maximum.
-  string id = 3 [ (google.api.field_behavior) = IMMUTABLE ];
+  string id = 3 [(google.api.field_behavior) = IMMUTABLE];
   // ConnectorDefinition title
-  string title = 4 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  string title = 4 [(google.api.field_behavior) = OUTPUT_ONLY];
   // ConnectorDefinition documentation URL
-  string documentation_url = 5 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  string documentation_url = 5 [(google.api.field_behavior) = OUTPUT_ONLY];
   // ConnectorDefinition icon
-  string icon = 6 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  string icon = 6 [(google.api.field_behavior) = OUTPUT_ONLY];
   // ConnectorDefinition spec
-  Spec spec = 7 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  Spec spec = 7 [(google.api.field_behavior) = OUTPUT_ONLY];
   // Connector Type
-  ConnectorType connector_type = 8 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  ConnectorType connector_type = 8 [(google.api.field_behavior) = OUTPUT_ONLY];
   // ConnectorDefinition tombstone, i.e., if not set or false, the
   // configuration is active, or otherwise, if true, this configuration is
   // permanently off
-  bool tombstone = 9 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  bool tombstone = 9 [(google.api.field_behavior) = OUTPUT_ONLY];
   // ConnectorDefinition public flag, i.e., true if this connector
   // definition is available to all workspaces
-  bool public = 10 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  bool public = 10 [(google.api.field_behavior) = OUTPUT_ONLY];
   // ConnectorDefinition custom flag, i.e., whether this is a custom
   // connector definition
-  bool custom = 11 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  bool custom = 11 [(google.api.field_behavior) = OUTPUT_ONLY];
   // ConnectorDefinition iconUrl
-  string icon_url = 12 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  string icon_url = 12 [(google.api.field_behavior) = OUTPUT_ONLY];
   // ConnectorDefinition vendor name
-  string vendor = 13 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  string vendor = 13 [(google.api.field_behavior) = OUTPUT_ONLY];
   // ConnectorDefinition vendorAttributes, i.e. the vendor-specific attributes
-  google.protobuf.Struct vendor_attributes = 14 [ (google.api.field_behavior) = OUTPUT_ONLY ];
+  google.protobuf.Struct vendor_attributes = 14 [(google.api.field_behavior) = OUTPUT_ONLY];
 }
 
 ///////////////////////////////////////////////////////////////////////
@@ -96,13 +99,13 @@ message ListConnectorDefinitionsRequest {
   // service may return fewer than this value. If unspecified, at most 10
   // ConnectorDefinitions will be returned. The maximum value is 100;
   // values above 100 will be coerced to 100.
-  optional int64 page_size = 1 [ (google.api.field_behavior) = OPTIONAL ];
+  optional int64 page_size = 1 [(google.api.field_behavior) = OPTIONAL];
   // Page token
-  optional string page_token = 2 [ (google.api.field_behavior) = OPTIONAL ];
+  optional string page_token = 2 [(google.api.field_behavior) = OPTIONAL];
   // Definition view (default is DEFINITION_VIEW_BASIC)
-  optional View view = 3 [ (google.api.field_behavior) = OPTIONAL ];
+  optional View view = 3 [(google.api.field_behavior) = OPTIONAL];
   // Filter expression to list connector definitions
-  optional string filter = 4 [ (google.api.field_behavior) = OPTIONAL ];
+  optional string filter = 4 [(google.api.field_behavior) = OPTIONAL];
 }
 
 // ListConnectorDefinitionsResponse represents a response for a list
@@ -123,18 +126,14 @@ message GetConnectorDefinitionRequest {
   // "connector-definitions/*"
   string name = 1 [
     (google.api.field_behavior) = REQUIRED,
-    (google.api.resource_reference) = {
-      type : "api.instill.tech/ConnectorDefinition"
-    },
+    (google.api.resource_reference) = {type: "api.instill.tech/ConnectorDefinition"},
     (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) = {
-      field_configuration : {
-        path_param_name : "connector_definition.name"
-      }
+      field_configuration: {path_param_name: "connector_definition.name"}
     }
   ];
   // ConnectorDefinition resource view (default is
   // DEFINITION_VIEW_BASIC)
-  optional View view = 2 [ (google.api.field_behavior) = OPTIONAL ];
+  optional View view = 2 [(google.api.field_behavior) = OPTIONAL];
 }
 
 // GetConnectorDefinitionResponse represents a
