+++ integration-test/rest-data-connector.js
@@ -11,9 +11,9 @@ export function CheckCreate() {
 
     group("Connector API: Create destination connectors", () => {
 
-        // response
+        // end
         var httpDstConnector = {
-            "id": "response",
+            "id": "end-operator",
             "connector_definition_name": constant.dstDefRscName,
             "description": "HTTP source",
             "configuration": {},
@@ -22,22 +22,25 @@ export function CheckCreate() {
         var resDst = http.request(
             "POST",
             `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(httpDstConnector), constant.params)
+            JSON.stringify(httpDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
+
         check(resDst, {
             "POST /v1alpha/connectors response status for creating HTTP destination connector 201": (r) => r.status === 201,
             "POST /v1alpha/connectors response connector name": (r) => r.json().connector.name == `connectors/${httpDstConnector.id}`,
             "POST /v1alpha/connectors response connector uid": (r) => helper.isUUID(r.json().connector.uid),
-            "POST /v1alpha/connectors response connector connector_definition_name": (r) => r.json().connector.connector_definition_name === constant.dstDefRscName,
-            "POST /v1alpha/connectors response connector owner is UUID": (r) => helper.isValidOwner(r.json().connector.user),
+            "POST /v1alpha/connectors response connector connector_definition_name": (r) => r.json().connector.connector_definition_name === constant.dstDefRscName
         });
 
         check(http.request(
             "POST",
             `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(httpDstConnector), constant.params), {
+            JSON.stringify(httpDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        }), {
             "POST /v1alpha/connectors response duplicate HTTP destination connector status 409": (r) => r.status === 409
         });
-
 
         // destination-csv
         var csvDstConnector = {
@@ -45,14 +48,14 @@ export function CheckCreate() {
             "connector_definition_name": constant.csvDstDefRscName,
             "description": randomString(50),
             "configuration": constant.csvDstConfig
+
         }
 
         var resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
 
-        check(resCSVDst, {
-            "POST /v1alpha/connectors response status 201": (r) => r.status === 201,
-        });
         http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
             {}, constant.params)
 
@@ -60,9 +63,41 @@ export function CheckCreate() {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
         })
 
+        check(resCSVDst, {
+            "POST /v1alpha/connectors response status 201": (r) => r.status === 201,
+        });
+
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id} response STATE_CONNECTED`]: (r) => r.json().connector.state === "STATE_CONNECTED",
         });
+
+        // destination-mysql (will end up with STATE_ERROR)
+        var mySQLDstConnector = {
+            "id": randomString(10),
+            "connector_definition_name": constant.mySQLDstDefRscName,
+            "configuration": {
+                "host": randomString(10),
+                "port": 3306,
+                "username": randomString(10),
+                "database": randomString(10),
+            }
+
+        }
+
+        var resDstMySQL = http.request(
+            "POST",
+            `${connectorPublicHost}/v1alpha/connectors`,
+            JSON.stringify(mySQLDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
+
+        check(resDstMySQL, {
+            "POST /v1alpha/connectors response status for creating MySQL destination connector 201": (r) => r.status === 201,
+            "POST /v1alpha/connectors response connector name": (r) => r.json().connector.name == `connectors/${mySQLDstConnector.id}`,
+            "POST /v1alpha/connectors response connector uid": (r) => helper.isUUID(r.json().connector.uid),
+            "POST /v1alpha/connectors response connector connector_definition_name": (r) => r.json().connector.connector_definition_name === constant.mySQLDstDefRscName
+        });
+
 
         // Delete test records
         check(http.request("DELETE", `${connectorPublicHost}/v1alpha/connectors/${resDst.json().connector.id}`), {
@@ -71,7 +106,9 @@ export function CheckCreate() {
         check(http.request("DELETE", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}`), {
             [`DELETE /v1alpha/connectors/${resCSVDst.json().connector.id} response status 204`]: (r) => r.status === 204,
         });
-
+        check(http.request("DELETE", `${connectorPublicHost}/v1alpha/connectors/${resDstMySQL.json().connector.id}`), {
+            [`DELETE /v1alpha/connectors/${resDstMySQL.json().connector.id} response status 204`]: (r) => r.status === 204,
+        });
     });
 
 }
@@ -80,7 +117,7 @@ export function CheckList() {
 
     group("Connector API: List destination connectors", () => {
 
-        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DESTINATION`), {
+        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DATA`), {
             [`GET /v1alpha/connectors response status is 200`]: (r) => r.status === 200,
             [`GET /v1alpha/connectors response connectors array is 0 length`]: (r) => r.json().connectors.length === 0,
             [`GET /v1alpha/connectors response next_page_token is empty`]: (r) => r.json().next_page_token === "",
@@ -101,55 +138,54 @@ export function CheckList() {
         // Create connectors
         for (const reqBody of reqBodies) {
             var resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-                JSON.stringify(reqBody), constant.params)
+                JSON.stringify(reqBody), {
+                headers: { "Content-Type": "application/json" },
+            })
             check(resCSVDst, {
                 [`POST /v1alpha/connectors x${reqBodies.length} response status 201`]: (r) => r.status === 201,
             });
         }
 
-        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DESTINATION`), {
+        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DATA`), {
             [`GET /v1alpha/connectors response status is 200`]: (r) => r.status === 200,
             [`GET /v1alpha/connectors response has connectors array`]: (r) => Array.isArray(r.json().connectors),
             [`GET /v1alpha/connectors response has total_size = ${numConnectors}`]: (r) => r.json().total_size == numConnectors,
         });
 
-        var limitedRecords = http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DESTINATION`)
-        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DESTINATION&page_size=0`), {
+        var limitedRecords = http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DATA`)
+        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DATA&page_size=0`), {
             "GET /v1alpha/connectors?page_size=0 response status is 200": (r) => r.status === 200,
             "GET /v1alpha/connectors?page_size=0 response all records": (r) => r.json().connectors.length === limitedRecords.json().connectors.length,
         });
 
-        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DESTINATION&page_size=1`), {
+        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DATA&page_size=1`), {
             "GET /v1alpha/connectors?page_size=1 response status is 200": (r) => r.status === 200,
             "GET /v1alpha/connectors?page_size=1 response connectors size 1": (r) => r.json().connectors.length === 1,
         });
 
-        var pageRes = http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DESTINATION&page_size=1`)
-        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?page_size=1&page_token=${pageRes.json().next_page_token}`), {
+        var pageRes = http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DATA&page_size=1`)
+        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DATA&page_size=1&page_token=${pageRes.json().next_page_token}`), {
             [`GET /v1alpha/connectors?page_size=1&page_token=${pageRes.json().next_page_token} response status is 200`]: (r) => r.status === 200,
             [`GET /v1alpha/connectors?page_size=1&page_token=${pageRes.json().next_page_token} response connectors size 1`]: (r) => r.json().connectors.length === 1,
         });
 
-        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DESTINATION&page_size=1&view=VIEW_BASIC`), {
+        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DATA&page_size=1&view=VIEW_BASIC`), {
             "GET /v1alpha/connectors?page_size=1&view=VIEW_BASIC response status 200": (r) => r.status === 200,
             "GET /v1alpha/connectors?page_size=1&view=VIEW_BASIC response connectors[0].configuration is null": (r) => r.json().connectors[0].configuration === null,
-            "GET /v1alpha/connectors?page_size=1&view=VIEW_BASIC response connectors[0].owner is UUID": (r) => helper.isValidOwner(r.json().connectors[0].user),
         });
 
-        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DESTINATION&page_size=1&view=VIEW_FULL`), {
+        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DATA&page_size=1&view=VIEW_FULL`), {
             "GET /v1alpha/connectors?page_size=1&view=VIEW_FULL response status 200": (r) => r.status === 200,
             "GET /v1alpha/connectors?page_size=1&view=VIEW_FULL response connectors[0].configuration is not null": (r) => r.json().connectors[0].configuration !== null,
             "GET /v1alpha/connectors?page_size=1&view=VIEW_FULL response connectors[0].connector_definition_detail is not null": (r) => r.json().connectors[0].connector_definition_detail !== null,
-            "GET /v1alpha/connectors?page_size=1&view=VIEW_FULL response connectors[0].owner is UUID": (r) => helper.isValidOwner(r.json().connectors[0].user),
         });
 
-        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DESTINATION&page_size=1`), {
+        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DATA&page_size=1`), {
             "GET /v1alpha/connectors?page_size=1 response status 200": (r) => r.status === 200,
             "GET /v1alpha/connectors?page_size=1 response connectors[0].configuration is null": (r) => r.json().connectors[0].configuration === null,
-            "GET /v1alpha/connectors?page_size=1 response connectors[0].owner is UUID": (r) => helper.isValidOwner(r.json().connectors[0].user),
         });
 
-        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DESTINATION&page_size=${limitedRecords.json().total_size}`), {
+        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors?filter=connector_type=CONNECTOR_TYPE_DATA&page_size=${limitedRecords.json().total_size}`), {
             [`GET /v1alpha/connectors?page_size=${limitedRecords.json().total_size} response status 200`]: (r) => r.status === 200,
             [`GET /v1alpha/connectors?page_size=${limitedRecords.json().total_size} response next_page_token is empty`]: (r) => r.json().next_page_token === ""
         });
@@ -172,25 +208,22 @@ export function CheckGet() {
             "connector_definition_name": constant.csvDstDefRscName,
             "description": randomString(50),
             "configuration": constant.csvDstConfig
-
         }
 
         var resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
 
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
+            [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
+        })
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id} response status 200`]: (r) => r.status === 200,
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id} response connector id`]: (r) => r.json().connector.id === csvDstConnector.id,
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id} response connector connector_definition_name permalink`]: (r) => r.json().connector.connector_definition_name === constant.csvDstDefRscName,
-            [`GET /v1alpha/connectors/${resCSVDst.json().connector.id} response connector owner is UUID`]: (r) => helper.isValidOwner(r.json().connector.user),
         });
-
-        check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
-            [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
-        })
 
         check(http.request("DELETE", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}`), {
             [`DELETE /v1alpha/connectors/${resCSVDst.json().connector.id} response status 204`]: (r) => r.status === 204,
@@ -207,11 +240,12 @@ export function CheckUpdate() {
             "connector_definition_name": constant.csvDstDefRscName,
             "description": randomString(50),
             "configuration": constant.csvDstConfig
-
         }
 
         var resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         var csvDstConnectorUpdate = {
             "id": csvDstConnector.id,
@@ -224,42 +258,44 @@ export function CheckUpdate() {
         }
 
         var resCSVDstUpdate = http.request("PATCH", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}`,
-            JSON.stringify(csvDstConnectorUpdate), constant.params)
+            JSON.stringify(csvDstConnectorUpdate), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         check(resCSVDstUpdate, {
             [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} response status 200`]: (r) => r.status === 200,
             [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} response connector id`]: (r) => r.json().connector.id === csvDstConnectorUpdate.id,
             [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} response connector connector_definition_name`]: (r) => r.json().connector.connector_definition_name === constant.csvDstDefRscName,
-            [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} response connector description`]: (r) => r.json().connector.description === csvDstConnectorUpdate.description,
+            [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} response connector description`]: (r) => r.json().connector.description === csvDstConnectorUpdate.connector.description,
             [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} response connector tombstone`]: (r) => r.json().connector.tombstone === false,
-            [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} response connector configuration`]: (r) => r.json().connector.configuration.destination_path === csvDstConnectorUpdate.configuration.destination_path,
-            [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} response connector owner is UUID`]: (r) => helper.isValidOwner(r.json().connector.user),
+            [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} response connector configuration`]: (r) => r.json().connector.configuration.destination_path === csvDstConnectorUpdate.connector.configuration.destination_path
         });
 
         // Try to update with empty description
         csvDstConnectorUpdate = {
-
-            "description": "",
-
+            "configuration": {}
         }
 
         resCSVDstUpdate = http.request("PATCH", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}`,
-            JSON.stringify(csvDstConnectorUpdate), constant.params)
+            JSON.stringify(csvDstConnectorUpdate), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         check(resCSVDstUpdate, {
             [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} with empty description response status 200`]: (r) => r.status === 200,
-            [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} with empty description response empty description`]: (r) => r.json().connector.description === csvDstConnectorUpdate.description,
+            [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} with empty description response empty description`]: (r) => r.json().connector.description === csvDstConnectorUpdate.connector.description,
         })
 
         // Try to update with a non-existing name field (which should be ignored because name field is OUTPUT_ONLY)
         csvDstConnectorUpdate = {
             "name": `connectors/${randomString(5)}`,
             "description": randomString(50),
-
         }
 
         resCSVDstUpdate = http.request("PATCH", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}`,
-            JSON.stringify(csvDstConnectorUpdate), constant.params)
+            JSON.stringify(csvDstConnectorUpdate), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         check(resCSVDstUpdate, {
             [`PATCH /v1alpha/connectors/${resCSVDst.json().connector.id} with non-existing name field response status 200`]: (r) => r.status === 200,
@@ -283,95 +319,19 @@ export function CheckLookUp() {
         }
 
         var resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.uid}/lookUp`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.uid}/lookUp response status 200`]: (r) => r.status === 200,
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.uid}/lookUp response connector uid`]: (r) => r.json().connector.uid === resCSVDst.json().connector.uid,
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.uid}/lookUp response connector connector_definition_name`]: (r) => r.json().connector.connector_definition_name === constant.csvDstDefRscName,
-            [`GET /v1alpha/connectors/${resCSVDst.json().connector.uid}/lookUp response connector owner is UUID`]: (r) => helper.isValidOwner(r.json().connector.user),
         });
 
         check(http.request("DELETE", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}`), {
             [`DELETE /v1alpha/connectors/${resCSVDst.json().connector.id} response status 204`]: (r) => r.status === 204,
         });
-
-    });
-}
-
-export function CheckConnect() {
-    group("Connector API: Check Connect", () => {
-
-
-        var csvDstConnector = {
-            "id": randomString(10),
-            "connector_definition_name": constant.csvDstDefRscName,
-            "description": randomString(50),
-            "configuration": {}
-        }
-
-        // Cannot connect with unfinished config
-        var resDstCsv = http.request(
-            "POST",
-            `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-
-        check(resDstCsv, {
-            "POST /v1alpha/connectors response status for creating MySQL destination connector 201": (r) => r.status === 201,
-            "POST /v1alpha/connectors response connector name": (r) => r.json().connector.name == `connectors/${csvDstConnector.id}`,
-            "POST /v1alpha/connectors response connector uid": (r) => helper.isUUID(r.json().connector.uid),
-            "POST /v1alpha/connectors response connector connector_definition_name": (r) => r.json().connector.connector_definition_name === constant.csvDstDefRscName,
-            "POST /v1alpha/connectors response connector owner is UUID": (r) => helper.isValidOwner(r.json().connector.user),
-        });
-
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resDstCsv.json().connector.id}/connect`), {
-            [`POST /v1alpha/connectors/${resDstCsv.json().connector.id}/connect response status 400`]: (r) => r.status === 400,
-        });
-
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resDstCsv.json().connector.id}/disconnect`), {
-            [`POST /v1alpha/connectors/${resDstCsv.json().connector.id}/disconnect response status 200`]: (r) => r.status === 200,
-        });
-
-        check(http.request("DELETE", `${connectorPublicHost}/v1alpha/connectors/${resDstCsv.json().connector.id}`), {
-            [`DELETE /v1alpha/connectors/${resDstCsv.json().connector.id} response status 204`]: (r) => r.status === 204,
-        });
-
-        var csvDstConnector = {
-            "id": randomString(10),
-            "connector_definition_name": constant.csvDstDefRscName,
-            "description": randomString(50),
-            "configuration": constant.csvDstConfig
-        }
-
-        // Cannot connect with unfinished config
-        var resDstCsv = http.request(
-            "POST",
-            `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-
-        check(resDstCsv, {
-            "POST /v1alpha/connectors response status for creating MySQL destination connector 201": (r) => r.status === 201,
-            "POST /v1alpha/connectors response connector name": (r) => r.json().connector.name == `connectors/${csvDstConnector.id}`,
-            "POST /v1alpha/connectors response connector uid": (r) => helper.isUUID(r.json().connector.uid),
-            "POST /v1alpha/connectors response connector connector_definition_name": (r) => r.json().connector.connector_definition_name === constant.csvDstDefRscName,
-            "POST /v1alpha/connectors response connector owner is UUID": (r) => helper.isValidOwner(r.json().connector.user),
-        });
-
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resDstCsv.json().connector.id}/connect`), {
-            [`POST /v1alpha/connectors/${resDstCsv.json().connector.id}/connect response status 200`]: (r) => r.status === 200,
-        });
-
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resDstCsv.json().connector.id}/disconnect`), {
-            [`POST /v1alpha/connectors/${resDstCsv.json().connector.id}/disconnect response status 200`]: (r) => r.status === 200,
-        });
-
-        check(http.request("DELETE", `${connectorPublicHost}/v1alpha/connectors/${resDstCsv.json().connector.id}`), {
-            [`DELETE /v1alpha/connectors/${resDstCsv.json().connector.id} response status 204`]: (r) => r.status === 204,
-        });
-
-
-
-
 
     });
 }
@@ -384,41 +344,56 @@ export function CheckState() {
             "connector_definition_name": constant.csvDstDefRscName,
             "description": randomString(50),
             "configuration": constant.csvDstConfig
-
         }
 
         var resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
+
+        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/disconnect`, null, {
+            headers: { "Content-Type": "application/json" }
+        }), {
+            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/disconnect response at UNSPECIFIED state status 200`]: (r) => r.status === 200,
+        });
+
+        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/connect`, null, {
+            headers: { "Content-Type": "application/json" }
+        }), {
+            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/connect response at UNSPECIFIED state status 200`]: (r) => r.status === 200,
+        });
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
         })
 
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/connect`, null, constant.params), {
+        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/connect`, null, {
+            headers: { "Content-Type": "application/json" }
+        }), {
             [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/connect response status 200 (with STATE_CONNECTED)`]: (r) => r.status === 200,
         });
 
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/disconnect`, null, constant.params), {
+        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/disconnect`, null, {
+            headers: { "Content-Type": "application/json" }
+        }), {
             [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/disconnect response status 200 (with STATE_CONNECTED)`]: (r) => r.status === 200,
         });
 
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/disconnect`, null, constant.params), {
+        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/disconnect`, null, {
+            headers: { "Content-Type": "application/json" }
+        }), {
             [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/disconnect response status 200 (with STATE_DISCONNECTED)`]: (r) => r.status === 200,
         });
 
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/connect`, null, constant.params), {
+        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/connect`, null, {
+            headers: { "Content-Type": "application/json" }
+        }), {
             [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/connect response status 200 (with STATE_DISCONNECTED)`]: (r) => r.status === 200,
         });
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
         })
-
-        check(http.request("DELETE", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}`), {
-            [`DELETE /v1alpha/connectors/${resCSVDst.json().connector.id} response status 204`]: (r) => r.status === 204,
-        });
 
     });
 
@@ -433,16 +408,19 @@ export function CheckRename() {
             "connector_definition_name": constant.csvDstDefRscName,
             "description": randomString(50),
             "configuration": constant.csvDstConfig
-
         }
 
         var resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/rename`,
             JSON.stringify({
                 "new_connector_id": `some-id-not-${resCSVDst.json().connector.id}`
-            }), constant.params), {
+            }), {
+            headers: { "Content-Type": "application/json" }
+        }), {
             [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/rename response status 200`]: (r) => r.status === 200,
             [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/rename response id is some-id-not-${resCSVDst.json().connector.id}`]: (r) => r.json().connector.id === `some-id-not-${resCSVDst.json().connector.id}`,
         });
@@ -470,10 +448,10 @@ export function CheckExecute() {
         }
 
         resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
+        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`, {})
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
@@ -507,9 +485,10 @@ export function CheckExecute() {
         }
 
         resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
+        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`, {})
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
@@ -543,9 +522,10 @@ export function CheckExecute() {
         }
 
         resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
+        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`, {})
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
@@ -575,12 +555,15 @@ export function CheckExecute() {
             "configuration": {
                 "destination_path": "/local/test-keypoint"
             },
+
         }
 
         resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
+        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`, {})
+
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
         })
@@ -609,12 +592,13 @@ export function CheckExecute() {
             "configuration": {
                 "destination_path": "/local/test-ocr"
             },
+
         }
 
         resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
@@ -643,26 +627,25 @@ export function CheckExecute() {
             "description": randomString(50),
             "configuration": {
                 "destination_path": "/local/test-semantic-segmentation"
-            }
+            },
         }
 
         resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
         })
 
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/execute`,
+        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/write`,
             JSON.stringify({
-                "inputs": constant.semanticSegModelOutputs
+                "inputs": constant.instSegModelOutputs
             }), {
             headers: { "Content-Type": "application/json" }
         }), {
-            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/execute response status 200 (semantic-segmentation)`]: (r) => r.status === 200,
+            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/write response status 200 (semantic-segmentation)`]: (r) => r.status === 200,
         });
 
         // Wait for 1 sec for the connector writing to the destination-csv before deleting it
@@ -680,25 +663,24 @@ export function CheckExecute() {
             "configuration": {
                 "destination_path": "/local/test-instance-segmentation"
             },
-
         }
 
         resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
         })
 
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/execute`,
+        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/write`,
             JSON.stringify({
                 "inputs": constant.instSegModelOutputs
             }), {
             headers: { "Content-Type": "application/json" }
         }), {
-            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/execute response status 200 (instance-segmentation)`]: (r) => r.status === 200,
+            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/write response status 200 (instance-segmentation)`]: (r) => r.status === 200,
         });
 
         // Wait for 1 sec for the connector writing to the destination-csv before deleting it
@@ -715,25 +697,25 @@ export function CheckExecute() {
             "description": randomString(50),
             "configuration": {
                 "destination_path": "/local/test-text-to-image"
-            }
+            },
         }
 
         resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
         })
 
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/execute`,
+        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/write`,
             JSON.stringify({
-                "inputs": constant.textToImageModelOutputs
+                "inputs": constant.textGenerationModelOutputs
             }), {
             headers: { "Content-Type": "application/json" }
         }), {
-            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/execute response status 200 (text-to-image)`]: (r) => r.status === 200,
+            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/write response status 200 (text-to-image)`]: (r) => r.status === 200,
         });
 
         // Wait for 1 sec for the connector writing to the destination-csv before deleting it
@@ -754,21 +736,21 @@ export function CheckExecute() {
         }
 
         resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
         })
 
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/execute`,
+        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/write`,
             JSON.stringify({
                 "inputs": constant.textGenerationModelOutputs
             }), {
             headers: { "Content-Type": "application/json" }
         }), {
-            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/execute response status 200 (text-generation)`]: (r) => r.status === 200,
+            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/write response status 200 (text-generation)`]: (r) => r.status === 200,
         });
 
         // Wait for 1 sec for the connector writing to the destination-csv before deleting it
@@ -785,25 +767,25 @@ export function CheckExecute() {
             "description": randomString(50),
             "configuration": {
                 "destination_path": "/local/test-unspecified"
-            }
+            },
         }
 
         resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
+            JSON.stringify(csvDstConnector), {
+            headers: { "Content-Type": "application/json" },
+        })
 
         check(http.request("GET", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/watch`), {
             [`GET /v1alpha/connectors/${resCSVDst.json().connector.id}/watch response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
         })
 
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/execute`,
+        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/write`,
             JSON.stringify({
                 "inputs": constant.unspecifiedModelOutputs
             }), {
             headers: { "Content-Type": "application/json" }
         }), {
-            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/execute response status 200 (unspecified)`]: (r) => r.status === 200,
+            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/write response status 200 (unspecified)`]: (r) => r.status === 200,
         });
 
         // Wait for 1 sec for the connector writing to the destination-csv before deleting it
@@ -811,34 +793,6 @@ export function CheckExecute() {
 
         check(http.request("DELETE", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}`), {
             [`DELETE /v1alpha/connectors/${resCSVDst.json().connector.id} response status 204 (unspecified)`]: (r) => r.status === 204,
-        });
-    });
-}
-
-export function CheckTest() {
-
-    group("Connector API: Test destination connectors by ID", () => {
-
-        var csvDstConnector = {
-            "id": randomString(10),
-            "connector_definition_name": constant.csvDstDefRscName,
-            "description": randomString(50),
-            "configuration": constant.csvDstConfig
-        }
-
-        var resCSVDst = http.request("POST", `${connectorPublicHost}/v1alpha/connectors`,
-            JSON.stringify(csvDstConnector), constant.params)
-
-        http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${csvDstConnector.id}/connect`,
-            {}, constant.params)
-
-        check(http.request("POST", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}/testConnection`), {
-            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/testConnection response status 200`]: (r) => r.status === 200,
-            [`POST /v1alpha/connectors/${resCSVDst.json().connector.id}/testConnection response connector state is STATE_CONNECTED`]: (r) => r.json().state === "STATE_CONNECTED",
-        });
-
-        check(http.request("DELETE", `${connectorPublicHost}/v1alpha/connectors/${resCSVDst.json().connector.id}`), {
-            [`DELETE /v1alpha/connectors/${resCSVDst.json().connector.id} response status 204`]: (r) => r.status === 204,
         });
     });
 }
