+++ pkg/constant/constant.go
@@ -4,5 +4,5 @@ package constant
 const DefaultOwnerID string = "instill-ai"
 const HeaderOwnerUIDKey = "jwt-sub"
 const HeaderOwnerIDKey = "owner-id"
-const TriggerConnectorId = "trigger"
-const ResponseConnectorId = "response"
+const StartConnectorId = "start-operator"
+const EndConnectorId = "end-operator"
