+++ pkg/usage/usage.go
@@ -13,8 +13,7 @@ import (
 	"github.com/instill-ai/x/repo"
 	"go.einride.tech/aip/filtering"
 
-	connectorDestination "github.com/instill-ai/connector-destination/pkg"
-	connectorSource "github.com/instill-ai/connector-source/pkg"
+	connectorData "github.com/instill-ai/connector-data/pkg"
 	connectorBase "github.com/instill-ai/connector/pkg/base"
 	mgmtPB "github.com/instill-ai/protogen-go/base/mgmt/v1alpha"
 	usagePB "github.com/instill-ai/protogen-go/base/usage/v1alpha"
@@ -35,8 +34,7 @@ type usage struct {
 	mgmtPrivateServiceClient mgmtPB.MgmtPrivateServiceClient
 	reporter                 usageReporter.Reporter
 	version                  string
-	connectorSource          connectorBase.IConnector
-	connectorDestination     connectorBase.IConnector
+	connectorData            connectorBase.IConnector
 }
 
 // NewUsage initiates a usage instance
@@ -60,8 +58,7 @@ func NewUsage(ctx context.Context, r repository.Repository, ma mgmtPB.MgmtPrivat
 		mgmtPrivateServiceClient: ma,
 		reporter:                 reporter,
 		version:                  version,
-		connectorDestination:     connectorDestination.Init(logger, connector.GetConnectorDestinationOptions()),
-		connectorSource:          connectorSource.Init(logger),
+		connectorData:            connectorData.Init(logger, connector.GetConnectorDataOptions()),
 	}
 }
 
@@ -87,21 +84,8 @@ func (u *usage) RetrieveUsageData() interface{} {
 		logger.Error(fmt.Sprintf("%s", err))
 	}
 
-	var srcParser filtering.Parser
-	srcParser.Init("connector_type=CONNECTOR_TYPE_SOURCE")
-	srcParsedExpr, err := srcParser.Parse()
-	if err != nil {
-		logger.Error(fmt.Sprintf("%s", err))
-	}
-	var srcChecker filtering.Checker
-	srcChecker.Init(srcParsedExpr.Expr, srcParsedExpr.SourceInfo, declarations)
-	srcCheckedExpr, err := srcChecker.Check()
-	if err != nil {
-		logger.Error(fmt.Sprintf("%s", err))
-	}
-
 	var dstParser filtering.Parser
-	dstParser.Init("connector_type=CONNECTOR_TYPE_DESTINATION")
+	dstParser.Init("connector_type=CONNECTOR_TYPE_DATA")
 	dstParsedExpr, err := dstParser.Parse()
 	if err != nil {
 		logger.Error(fmt.Sprintf("%s", err))
@@ -127,54 +111,6 @@ func (u *usage) RetrieveUsageData() interface{} {
 		for _, user := range userResp.Users {
 
 			connPageToken := ""
-
-			srcConnConnectedStateNum := int64(0)
-			srcConnDisconnectedStateNum := int64(0)
-			srcConnDefSet := make(map[string]struct{})
-			for {
-				dbSrcConns, _, connNextPageToken, err := u.repository.ListConnectors(
-					ctx,
-					fmt.Sprintf("users/%s", user.GetUid()),
-					int64(repository.MaxPageSize),
-					connPageToken,
-					true,
-					filtering.Filter{
-						CheckedExpr: srcCheckedExpr,
-					},
-				)
-
-				if err != nil {
-					logger.Error(fmt.Sprintf("%s", err))
-				}
-
-				for _, conn := range dbSrcConns {
-					if conn.State == datamodel.ConnectorState(connectorPB.Connector_STATE_CONNECTED) {
-						srcConnConnectedStateNum++
-					}
-					if conn.State == datamodel.ConnectorState(connectorPB.Connector_STATE_DISCONNECTED) {
-						srcConnDisconnectedStateNum++
-					}
-					srcConnDef, err := u.connectorSource.GetConnectorDefinitionByUid(conn.ConnectorDefinitionUID)
-					if err != nil {
-						logger.Error(fmt.Sprintf("%s", err))
-					}
-					srcConnDefSet[srcConnDef.GetId()] = struct{}{}
-				}
-
-				if connNextPageToken == "" {
-					break
-				} else {
-					connPageToken = connNextPageToken
-				}
-			}
-
-			srcConnDefs := make([]string, 0, len(srcConnDefSet))
-			for k := range srcConnDefSet {
-				srcConnDefs = append(srcConnDefs, k)
-			}
-
-			connPageToken = ""
-
 			dstConnConnectedStateNum := int64(0)
 			dstConnDisconnectedStateNum := int64(0)
 			dstConnDefSet := make(map[string]struct{})
@@ -201,7 +137,7 @@ func (u *usage) RetrieveUsageData() interface{} {
 					if conn.State == datamodel.ConnectorState(connectorPB.Connector_STATE_DISCONNECTED) {
 						dstConnDisconnectedStateNum++
 					}
-					dstConnDef, err := u.connectorDestination.GetConnectorDefinitionByUid(conn.ConnectorDefinitionUID)
+					dstConnDef, err := u.connectorData.GetConnectorDefinitionByUid(conn.ConnectorDefinitionUID)
 					if err != nil {
 						logger.Error(fmt.Sprintf("%s", err))
 					}
@@ -222,9 +158,6 @@ func (u *usage) RetrieveUsageData() interface{} {
 
 			pbConnectorUsageData = append(pbConnectorUsageData, &usagePB.ConnectorUsageData_UserUsageData{
 				UserUid:                                  user.GetUid(),
-				SourceConnectorConnectedStateNum:         srcConnConnectedStateNum,
-				SourceConnectorDisconnectedStateNum:      srcConnDisconnectedStateNum,
-				SourceConnectorDefinitionIds:             srcConnDefs,
 				DestinationConnectorConnectedStateNum:    dstConnConnectedStateNum,
 				DestinationConnectorDisconnectedStateNum: dstConnDisconnectedStateNum,
 				DestinationConnectorDefinitionIds:        dstConnDefs,
